Trocando variável de 'int' para 'double'
Alteração nas variáveis para calcular de forma correta a porcentagem de bonificação.

+++ Professor.cs
@@ -46,7 +46,7 @@ namespace SistemaLMS
         {
             return this.ValorHora * horas;
         }
-        public double CalcularSalario(int horas, int bonificacao)
+        public double CalcularSalario(int horas, double bonificacao)
         {
             double valor = this.ValorHora * horas;
             double pct = bonificacao / 100;
